Update to website data sending

+++ include/BioreactorVaribiles.hpp
@@ -14,27 +14,26 @@ static constexpr double MU = 0.0006922; // Pa * s
 static constexpr double RHO = 993; // kg / m^3
 
 //Calculates flowrate
-double flowRateCalc(double sesnorValue) {
-    double fl = (((CHANNEL_WIDTH * CHANNEL_HEIGHT * CHANNEL_HEIGHT * sesnorValue) / (6 * MU)) * 6e7)/32;
+float flowRateCalc(float sesnorValue) {
+    float fl = (((CHANNEL_WIDTH * CHANNEL_HEIGHT * CHANNEL_HEIGHT * sesnorValue) / (6 * MU)) * 6e7)/32;
     return fl;
 }
 
 //Calculates the shear stress in Pa that will result from the provided flow rate in ml/min.
-double shearStressCalc(double flowRate) {
-    double sh = (flowRate / 6e7) * (6 * MU) / (CHANNEL_WIDTH * CHANNEL_HEIGHT * CHANNEL_HEIGHT);
+float shearStressCalc(float flowRate) {
+    float sh = (flowRate / 6e7) * (6 * MU) / (CHANNEL_WIDTH * CHANNEL_HEIGHT * CHANNEL_HEIGHT);
     return sh;
 }
 
 //Calculates the flow rate ml/min needed to generate specified chear stress in Pa
-double flowRateBasedOnShearStressCalc(double shearStress) {
-    double fl = (shearStress * 1e7 *  CHANNEL_WIDTH * CHANNEL_HEIGHT * CHANNEL_HEIGHT) / MU;
+float flowRateBasedOnShearStressCalc(float shearStress) {
+    float fl = (shearStress * 1e7 *  CHANNEL_WIDTH * CHANNEL_HEIGHT * CHANNEL_HEIGHT) / MU;
     return fl;
 }
 
-
 //Calculates the Reynolds number for the given flow rate.
-double reynoldsCalc(double flowRate) {
-    double r = (RHO * (flowRate / (CHANNEL_HEIGHT * CHANNEL_WIDTH)) * CHANNEL_HEIGHT) / MU;
+float reynoldsCalc(float flowRate) {
+    float r = (RHO * (flowRate / (CHANNEL_HEIGHT * CHANNEL_WIDTH)) * CHANNEL_HEIGHT) / MU;
     return r;
 }
